requirements.txt
uvicorn==0.24.0
streamlit
python-multipart==0.0.6
backoff
aiohttp
asyncio
openai-whisper
torch
requests
nest-asyncio
openai